# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/label-groups": {
      "get": {
        "summary": "Read Label Groups",
        "description": "Gets all label groups of the current user for a novel.",
        "operationId": "read_label_groups_label_groups_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Novelid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelGroup"
                  },
                  "title": "Response Read Label Groups Label Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Label Group",
        "description": "Creates a new label group.\n\nRaises:\n    404: Novel not found.\n    400: Label group name is too long.",
        "operationId": "create_label_group_label_groups_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelGroup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-groups/{labelGroupId}": {
      "get": {
        "summary": "Read Label Group",
        "description": "Gets a label group by id.\n\nRaises:\n    404: Label group not found (or insufficient permissions).",
        "operationId": "read_label_group_label_groups__labelGroupId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labelgroupid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelGroup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Label Group",
        "description": "Updates a label group (e.g. rename).\n\nRaises:\n    404: Label group not found.\n    400: Label group name is too long.",
        "operationId": "update_label_group_label_groups__labelGroupId__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labelgroupid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabelGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelGroup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-datas": {
      "get": {
        "summary": "Read Label Datas By Group Chapters",
        "description": "Gets all label datas in a label group, optionally filtered by chapter range.",
        "operationId": "read_label_datas_by_group_chapters_label_datas_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelGroupId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labelgroupid"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelData"
                  },
                  "title": "Response Read Label Datas By Group Chapters Label Datas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-datas/{labelDataId}": {
      "get": {
        "summary": "Read Label Data",
        "description": "Gets a label data by id.\n\nRaises:\n    404: Label data not found (or insufficient permissions).",
        "operationId": "read_label_data_label_datas__labelDataId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelDataId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labeldataid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Label Data Stream",
        "description": "Applies a stream of edit operations to labels.\n\nRaises:\n    404: Label data or its underlying revision text not found, or target label does not exist.\n    409: Word mismatch or label overlap detected.\n    400: Operation positions out of bounds or invalid operation.",
        "operationId": "update_label_data_stream_label_datas__labelDataId__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelDataId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labeldataid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabelDataStream"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-datas/{labelDataId}/labels": {
      "get": {
        "summary": "Read Labels By Label Data",
        "description": "Get the specific list of labels inside a label data entry.",
        "operationId": "read_labels_by_label_data_label_datas__labelDataId__labels_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelDataId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labeldataid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Label"
                  },
                  "title": "Response Read Labels By Label Data Label Datas  Labeldataid  Labels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-groups/{labelGroupId}/label-datas": {
      "post": {
        "summary": "Create Label Data",
        "description": "Creates a label data entry for a revision text in a label group.\n\nRaises:\n    404: Label group or revision text not found.\n    409: Label data for this revision text already exists in this group.",
        "operationId": "create_label_data_label_groups__labelGroupId__label_datas_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labelgroupid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/label-groups/{labelGroupId}/label-datas/auto-labels": {
      "post": {
        "summary": "Create Label Datas By Auto Labels",
        "description": "Creates label datas and populates labels from autolabel results.",
        "operationId": "create_label_datas_by_auto_labels_label_groups__labelGroupId__label_datas_auto_labels_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "labelGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Labelgroupid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelDataByAutoLabel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLabelDataByAutoLabelStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLabelData": {
        "properties": {
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          }
        },
        "type": "object",
        "required": [
          "chapterContentId"
        ],
        "title": "CreateLabelData",
        "description": "Pydantic schema for validating create requests for label data.\n\nAttributes:\n    chapter_content_id: UUID of chapter content being labelled."
      },
      "CreateLabelDataByAutoLabel": {
        "properties": {
          "modelName": {
            "type": "string",
            "title": "Modelname"
          },
          "modelParams": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 50,
            "title": "Modelparams"
          },
          "chapterIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapterids"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "modelName",
          "modelParams"
        ],
        "title": "CreateLabelDataByAutoLabel",
        "description": "Pydantic schema to specifiy a set of AutoLabels to be moved to LabelDatas.\n\nAttributes:\n    model_name: Name of NER model that performed the autolabeling.\n    model_params: Parameters of model used.\n    chapter_ids: Optional filter on what chapters to include.\n    start: Optional filter on the least chapter number to include.\n    end: Optional filter on the greatest chapter number to include."
      },
      "CreateLabelDataByAutoLabelStatus": {
        "properties": {
          "success": {
            "items": {
              "prefixItems": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string",
                  "format": "uuid"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Success"
          },
          "errors": {
            "items": {
              "prefixItems": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 3,
              "minItems": 3
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "success",
          "errors"
        ],
        "title": "CreateLabelDataByAutoLabelStatus",
        "description": "Return message for CreateLabelDataByAutoLabel.\n\nAttributes:\n    success: List of tuples of (chapter_id, chapter_content_id) for successful inserts.\n    errors: List of tuples of (chapter_id, chapter_content_id, error message) for failed inserts."
      },
      "CreateLabelGroup": {
        "properties": {
          "labelGroupName": {
            "type": "string",
            "maxLength": 31,
            "title": "Labelgroupname"
          },
          "novelId": {
            "type": "string",
            "format": "uuid",
            "title": "Novelid"
          }
        },
        "type": "object",
        "required": [
          "labelGroupName",
          "novelId"
        ],
        "title": "CreateLabelGroup",
        "description": "Pydantic schema for validating forms for creating a label group.\n\nAttributes:\n    label_group_name: Name of label group to create.\n    novel_id: UUID of novel this label group belongs to."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Label": {
        "properties": {
          "labelEntityGroup": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Labelentitygroup"
          },
          "labelScore": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Labelscore"
          },
          "labelWord": {
            "type": "string",
            "maxLength": 128,
            "title": "Labelword"
          },
          "labelStart": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Labelstart"
          },
          "labelEnd": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Labelend"
          },
          "labelDirty": {
            "type": "boolean",
            "title": "Labeldirty"
          },
          "labelDataId": {
            "type": "string",
            "format": "uuid",
            "title": "Labeldataid"
          },
          "labelId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelid"
          }
        },
        "type": "object",
        "required": [
          "labelEntityGroup",
          "labelScore",
          "labelWord",
          "labelStart",
          "labelEnd",
          "labelDirty",
          "labelDataId",
          "labelId"
        ],
        "title": "Label",
        "description": "Pydantic schema for a label that belongs to a LabelData.\nExtends LabelBase with a label_data_id foreign key and label_id pkey."
      },
...
      "LabelData": {
        "properties": {
          "labelDataId": {
            "type": "string",
            "format": "uuid",
            "title": "Labeldataid"
          },
          "labelGroupId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelgroupid"
          },
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          }
        },
        "type": "object",
        "required": [
          "labelDataId",
          "labelGroupId",
          "chapterContentId"
        ],
        "title": "LabelData",
        "description": "Pydantic schema for a list of labels in some text.\n\nAttributes:\n    label_data_id: UUID identifier for this LabelData.\n    label_group_id: UUID of label group this LabelData belongs to.\n    chapter_content_id: UUID of chapter content this LabelData is labelling."
      },
...
      "LabelGroup": {
        "properties": {
          "labelGroupId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelgroupid"
          },
          "labelGroupName": {
            "type": "string",
            "maxLength": 31,
            "title": "Labelgroupname"
          },
          "novelId": {
            "type": "string",
            "format": "uuid",
            "title": "Novelid"
          }
        },
        "type": "object",
        "required": [
          "labelGroupId",
          "labelGroupName",
          "novelId"
        ],
        "title": "LabelGroup",
        "description": "Pydantic schema for a label group.\n\nAttributes:\n    label_group_id: UUID identifier for this label group.\n    label_group_name: Name of this label group.\n    novel_id: UUID of novel this label group belongs to."
      },
...
      "UpdateLabelDataStream": {
        "properties": {
          "ops": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AddLabelOp"
                },
                {
                  "$ref": "#/components/schemas/DeleteLabelOp"
                },
                {
                  "$ref": "#/components/schemas/UpdateLabelOp"
                }
              ],
              "discriminator": {
                "propertyName": "op",
                "mapping": {
                  "add": "#/components/schemas/AddLabelOp",
                  "delete": "#/components/schemas/DeleteLabelOp",
                  "update": "#/components/schemas/UpdateLabelOp"
                }
              }
            },
            "type": "array",
            "title": "Ops"
          }
        },
        "type": "object",
        "required": [
          "ops"
        ],
        "title": "UpdateLabelDataStream",
        "description": "Pydantic schema for a buffered stream of label operations.\n\nAttributes:\n    ops: A list of label operations."
      },
      "UpdateLabelGroup": {
        "properties": {
          "labelGroupName": {
            "type": "string",
            "maxLength": 31,
            "title": "Labelgroupname"
          }
        },
        "type": "object",
        "required": [
          "labelGroupName"
        ],
        "title": "UpdateLabelGroup",
        "description": "Pydantic schema for validating forms for updating a label group.\n\nAttributes:\n    label_group_name: New name of label group."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}